+++ packages/terra-date-time-picker/src/terra-dev-site/test/date-time-picker/DateTimePickerMinMax.test.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import moment from 'moment';
+import DateTimePickerExampleTemplate from '../../common/DateTimePickerExampleTemplate';
+
+const DateTimePickerExample = () => (
+  <DateTimePickerExampleTemplate
+    minDate={moment().format()}
+    maxDate={moment().add(6, 'days').format()}
+    value="2019-04-19T10:30:00"
+  />
+);
+
+export default DateTimePickerExample;