+++ packages/terra-date-time-picker/tests/jest/DateTimePicker.test.jsx
@@ -37,6 +37,11 @@ it('should render a date time picker with included dates', () => {
   expect(datePicker).toMatchSnapshot();
 });
 
+it('should render a date time picker with min and max dates', () => {
+  const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" maxDate="2017-04-01" minDate="2017-04-10" utcOffset={0} />);
+  expect(datePicker).toMatchSnapshot();
+});
+
 it('should render a date time picker with releaseFocus', () => {
   const releaseFocus = () => {};
   const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" releaseFocus={releaseFocus} utcOffset={0} />);
@@ -71,9 +76,9 @@ it('should render a disabled date time picker', () => {
   expect(datePicker).toMatchSnapshot();
 });
 
-it('Should not throw any errors while date value is excluded and new time value is being entered ', () => {
+it('Should not throw any errors while date value is outside of the Min, Max range and new time value is being entered ', () => {
   const datePicker = mountWithIntl((
-    <DateTimePicker name="date-time-input" excludeDates={['2017-04-01']} />
+    <DateTimePicker name="date-time-input" maxDate="2017-04-01T10:30" minDate="2017-04-10T10:30" />
   ));
 
   const dateInput = datePicker.find({ name: 'terra-date-input', type: 'text' }).at(0);